pyproject.toml
[tool.poetry]
name = "django-training"
version = "0.1.0"
description = ""
authors = ["ibrahimramadan <ibrahim.ramadan.bld.ai>"]
readme = "README.md"
packages = [{include = "django_training"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.2"
environ = "^1.0"
django-environ = "^0.9.0"
django-imagekit = "^4.1.0"
Pillow = "^9.2.0"
djangorestframework = "^3.14.0"
django-filter = "^22.1"
django-extensions = "^3.2.1"
django-rest-knox = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
